old-tests/submission/Microsoft/windowobject/security_location_0.htm

<!DOCTYPE html>
<html>
    <head>
        <title>HTML 5 Test Suite: Window Object Security</title>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://www.w3.org/TR/html5/browsers.html" />
        <meta name="assert" content="access location object from different origins doesn't raise SECURITY_ERR exception" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
        <p>Access location object from different origins doesn't raise SECURITY_ERR exception</p>
        <div id=log></div>
        <script type="text/javascript">
            setup({explicit_done: true});
            function runTest() 
            {
              test(function()
              {
                // Add more tests upon passing increment the 'count'
                var ExpectedPasses = 1;
                var ActualPasses   = 0;
                ActualPasses = locationObjectTest();
                assert_equals(ActualPasses, ExpectedPasses);
              }, "access location object from different origins");
              done();
           }
           function locationObjectTest()
           {
                var frame = document.getElementById('testframe');

                try
                {
                    frame.setAttribute('onload', '');
                    frame.contentWindow.location = 'http://dev.w3.org/';

                    return 1;
                }
                catch(exc)
                {
                    return 0;
                }
            }
        </script>
        <iframe id='testframe' src="http://www.w3.org/" style="visibility: hidden; width: 0px; height:0px;" onload="runTest()">Test Frame</iframe>
    </body>
</html> 
